deleted service file and moved the methods in fileService,some minor fixes
