+++ src/app/core/services/file.service.ts
@@ -157,21 +157,38 @@ export class FileService {
   }
 
   getReceiptExtension(url: string) {
-    let res = null;
+    let receiptExtension = null;
     const name = url.split('?')[0];
     if (name) {
       const filename = name.toLowerCase();
-      const idx = filename.lastIndexOf('.');
+      const index = filename.lastIndexOf('.');
 
-      if (idx > -1) {
-        res = filename.substring(idx + 1, filename.length);
+      if (index > -1) {
+        receiptExtension = filename.substring(index + 1, filename.length);
       }
     }
 
-    return res;
+    return receiptExtension;
   }
 
   getImageTypeFromDataUrl(dataUrl: string) {
     return dataUrl.split(';')[0].split(':')[1];
+  }
+
+  getReceiptsDetails(file: FileObject) {
+    const receiptExtn = this.getReceiptExtension(file.name);
+    const receiptInfo = {
+      type: 'unknown',
+      thumbnail: 'img/fy-receipt.svg',
+    };
+
+    if (receiptExtn === 'pdf') {
+      receiptInfo.type = 'pdf';
+      receiptInfo.thumbnail = 'img/fy-pdf.svg';
+    } else if (receiptExtn && ['png', 'jpg', 'jpeg', 'gif'].indexOf(receiptExtn) > -1) {
+      receiptInfo.type = 'image';
+      receiptInfo.thumbnail = file.url;
+    }
+    return receiptInfo;
   }
 }
