+++ src/app/shared/components/summary-tile/summary-tile.component.html
@@ -39,17 +39,15 @@
   <div class="summary-tile--approval-state-block">
     <div class="summary-tile--approval-state-block-header">
       <div class="summary-tile--approvers">Approvers</div>
-      <div *ngIf="actions">
-        <div class="summary-tile--approvers-add-approver">
-          <app-fy-approver
-            *ngIf="actions.can_add_approver"
-            [id]="advanceRequest?.areq_id"
-            [ownerEmail]="advanceRequest?.us_email"
-            [type]="'ADVANCE_REQUEST'"
-            (notify)="onUpdateApprover($event)"
-          >
-          </app-fy-approver>
-        </div>
+      <div class="summary-tile--approvers-add-approver">
+        <app-fy-approver
+          *ngIf="actions.can_add_approver"
+          [id]="advanceRequest?.areq_id"
+          [ownerEmail]="advanceRequest?.us_email"
+          [type]="'ADVANCE_REQUEST'"
+          (notify)="onUpdateApprover($event)"
+        >
+        </app-fy-approver>
       </div>
     </div>
     <div *ngFor="let ap of approvals">
